修改有符号数与无符号的混用

+++ services/miscdevice_service/src/miscdevice_dump.cpp
@@ -62,7 +62,7 @@ void MiscdeviceDump::ParseCommand(int32_t fd, const std::vector<std::string>& ar
             continue;
         }
         if (str.find("-") == 0) {
-            count += str.size() - 1;
+            count += static_cast<int32_t>(str.size()) - 1;
             continue;
         }
     }
